Reorganize templates
