+++ backstage/.prettierignore
@@ -5,4 +5,4 @@ coverage
 
 # The Nunjucks templates for the Pull Requests use Nunjucks syntax, not YAML.
 # We can't rename the file using the `fetch:template` action so it is invalid YAML.
-templates/*/changes/*.yaml
+templates/*/pull-request-changes/*.yaml
