+++ backstage/app-config.yaml
@@ -104,6 +104,6 @@ catalog:
 
     # Resource Provisioner template
     - type: file
-      target: ../../templates/resource-provisioner.yaml
+      target: ../../templates/project-create/template.yaml
       rules:
         - allow: [Template]
