+++ backstage/templates/project-create/template.yaml
@@ -122,7 +122,7 @@ spec:
       name: Prepare Pull Request
       action: fetch:template
       input:
-        url: ./${{steps.ctx.output.template}}/changes
+        url: ./pull-request-changes
         values: ${{steps.ctx.output.template_values}}
 
     - id: debug-pull-request
